Refactored an RNN base class out of the lstm model which will be extended by the attention model as well
